requirements.txt
Django==6.0.2
djangorestframework==3.16.1
django-filter==25.2
Markdown==3.10.2
python-telegram-bot==22.6
python-dotenv==1.2.1
httpx==0.28.1